+++ R/nexus-internal.R
@@ -21,6 +21,27 @@ flatten_chr <- function(x, by) {
   z
 }
 
+#' Plotting Dimensions of Character Strings
+#'
+#' Convert string length in inch to number of (margin) lines.
+#' @param x A [`character`] vector of string whose length is to be calculated.
+#' @param ... Further parameter to be passed to [graphics::strwidth()]`, such as
+#'  `cex`.
+#' @return
+#'  A [`numeric`] vector (maximum string width in units of margin lines).
+#' @note For internal use only.
+#' @family graphic tools
+#' @keywords internal
+#' @noRd
+width2line <- function(x, ...) {
+  (max(graphics::strwidth(x, units = "inch", ...)) /
+     graphics::par("cin")[2] + graphics::par("mgp")[2]) * graphics::par("cex")
+}
+height2line <- function(x, ...) {
+  (max(graphics::strheight(x, units = "inch", ...)) /
+     graphics::par("cin")[2] + graphics::par("mgp")[2]) * graphics::par("cex")
+}
+
 #' Label Percentages
 #'
 #' @param x A [`numeric`] vector.