+++ R/split.R
@@ -30,3 +30,31 @@ split.LogRatio <- function(x, f, drop = FALSE, ...) {
 #' @rdname split
 #' @aliases split,LogRatio-method
 setMethod("split", "LogRatio", split.LogRatio)
+
+# Bind =========================================================================
+#' @export
+#' @rdname bind
+#' @aliases rbind2,CompositionMatrix,CompositionMatrix-method
+setMethod(
+  f = "rbind2",
+  signature = c(x = "CompositionMatrix", y = "CompositionMatrix"),
+  definition = function(x, y) {
+    mtx_x <- methods::as(x, "matrix")
+    mtx_y <- methods::as(y, "matrix")
+
+    spl <- c(rownames(x), rownames(y))
+    if (any(duplicated(spl))) {
+      warning("Duplicated rownames!", call. = FALSE)
+      spl <- make.unique(spl, sep = "_")
+    }
+
+    z <- rbind(mtx_x, mtx_y)
+    rownames(z) <- spl
+
+    .CompositionMatrix(
+      z,
+      totals = c(get_totals(x), get_totals(y)),
+      groups = c(get_groups(x), get_groups(y))
+    )
+  }
+)