+++ inst/examples/ex-barplot.R
@@ -5,13 +5,15 @@ data("hongite")
 coda <- as_composition(hongite)
 
 ## Bar plot
-barplot(coda, order = 2)
+barplot(coda)
 
 ## Data from Day et al. 2011
-data("kommos", package = "folio") # Coerce to compositional data
+data("kommos", package = "folio")
 kommos <- remove_NA(kommos, margin = 1) # Remove cases with missing values
-coda <- as_composition(kommos, groups = 1)
+coda <- as_composition(kommos, groups = 1) # Coerce to compositional data
 
 ## Use ceramic types for grouping
-barplot(coda, order = 1)
-barplot(coda, order = 1, horiz = FALSE)
+barplot(coda, order_columns = TRUE)
+
+## Display only minor elements
+barplot(coda, subset = element_minor(coda), order_columns = TRUE)